git-svn-id: http://svn.code.sf.net/p/dexer/code@45 805c3bd2-08bd-42b9-9c3f-09eed5005116

+++ Dexer/IO/InstructionWriter.cs
@@ -626,9 +626,9 @@ namespace Dexer.IO
                 {
                     case 1:
                         if (next)
-                            Codes[ExtraOffset++] |= (ushort)(Convert.ToSByte(element) << 8);
+                            Codes[ExtraOffset++] |= (ushort)((byte)(Convert.ToSByte(element)) << 8);
                         else
-                            Codes[ExtraOffset] |= (ushort)Convert.ToSByte(element);
+                            Codes[ExtraOffset] |= (ushort)((byte)Convert.ToSByte(element));
                         next = !next;
                         break;
                     case 2:
